chore: begin convert int long, ect -> int32_t, uin64_t, ect...

+++ ft_convert_base.c
@@ -6,7 +6,7 @@
 /*   By: sliziard <sliziard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/20 14:44:45 by sliziard          #+#    #+#             */
-/*   Updated: 2025/02/14 15:07:14 by sliziard         ###   ########.fr       */
+/*   Updated: 2025/04/08 22:04:57 by sliziard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include <limits.h>
 #include <stdlib.h>
 
-static bool	_is_in_base(char c, char *base, int *index)
+static bool	_is_in_base(char c, char *base, int32_t *index)
 {
-	int	i;
+	int32_t	i;
 
 	i = 0;
 	while (base[i])
@@ -31,13 +31,13 @@ static bool	_is_in_base(char c, char *base, int *index)
 	return (false);
 }
 
-int	ft_atoi_base(char *str, char *base)
+int32_t	ft_atoi_base(char *str, char *base)
 {
-	long	r;
+	int64_t	r;
 	size_t	b_len;
-	int		sign;
-	int		b_index;
-	int		i;
+	int32_t		sign;
+	int32_t		b_index;
+	int32_t		i;
 
 	if (!ft_valid_base(base, &b_len))
 		return (0);
@@ -60,7 +60,7 @@ int	ft_atoi_base(char *str, char *base)
 	return (r * sign);
 }
 
-char	*ft_itoa_base(int n, char *base)
+char	*ft_itoa_base(int32_t n, char *base)
 {
 	t_mem	*dest;
 	size_t	b_len;
@@ -68,16 +68,16 @@ char	*ft_itoa_base(int n, char *base)
 
 	if (!ft_valid_base(base, &b_len))
 		return (NULL);
-	dest = _init_dest_to_convert((long)n, b_len);
+	dest = _init_dest_to_convert((int64_t)n, b_len);
 	if (!dest)
 		return (NULL);
-	_nb_to_asciistr((long)n, dest, base, b_len);
+	_nb_to_asciistr((int64_t)n, dest, base, b_len);
 	n_in_b = dest->content;
 	free(dest);
 	return (n_in_b);
 }
 
-char	*ft_ltoa_base(long n, char *base)
+char	*ft_ltoa_base(int64_t n, char *base)
 {
 	t_mem	*dest;
 	size_t	b_len;
@@ -98,19 +98,19 @@ char	*ft_ltoa_base(long n, char *base)
 	return (n_in_b);
 }
 
-char	*ft_ulltoa_base(unsigned long long n, char *base)
+char	*ft_ulltoa_base(uint64_t n, char *base)
 {
-	unsigned long long	tmp;
-	size_t				base_len;
-	t_mem				dest;
+	uint64_t	tmp;
+	size_t		base_len;
+	t_mem		dest;
 
 	if (!ft_valid_base(base, &base_len))
 		return (NULL);
 	tmp = n;
 	dest.size = 1;
-	while (tmp >= (unsigned long long)base_len)
+	while (tmp >= (uint64_t)base_len)
 	{
-		tmp /= (unsigned long long)base_len;
+		tmp /= (uint64_t)base_len;
 		dest.size++;
 	}
 	dest.content = malloc(sizeof(char) * (dest.size + 1));
