+++ ft_is.c
@@ -6,32 +6,34 @@
 /*   By: sliziard <sliziard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 15:22:04 by sliziard          #+#    #+#             */
-/*   Updated: 2024/11/04 15:22:58 by sliziard         ###   ########.fr       */
+/*   Updated: 2025/04/08 22:11:10 by sliziard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_isalpha(int c)
+#include <stdint.h>
+
+int16_t	ft_isalpha(int32_t c)
 {
 	return ((c >= 'A' && c <= 'Z') || \
 		(c >= 'a' && c <= 'z'));
 }
 
-int	ft_isdigit(int c)
+int16_t	ft_isdigit(int32_t c)
 {
 	return (c >= '0' && c <= '9');
 }
 
-int	ft_isalnum(int c)
+int16_t	ft_isalnum(int32_t c)
 {
 	return (ft_isalpha(c) || ft_isdigit(c));
 }
 
-int	ft_isascii(int c)
+int16_t	ft_isascii(int32_t c)
 {
 	return (c >= 0 && c <= 127);
 }
 
-int	ft_isprint(int c)
+int16_t	ft_isprint(int32_t c)
 {
 	return (c >= 32 && c <= 126);
 }
